+++ CHANGELOG.md
@@ -3,3 +3,4 @@
 ## 0.0.4
 - Inital commits
 - Added Angular2 and dart2JS transformers, updated the cheatsheet doc to match
+- Added example of the Elvis operator from Alpha 26